Removed myownp2pn
