+++ src/transactions/propagating_the_tx.py
@@ -4,7 +4,7 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
-from node.myownp2pn import mynode
+from node.node import Node
 from node.unl import Unl
 
 
@@ -25,4 +25,4 @@ def PropagatingtheTX(tx):
         "transaction_time": tx.transaction_time,
     }
     for each_node in Unl.get_as_node_type(Unl.get_unl_nodes()):
-        mynode.main_node.send_data_to_node(each_node, items)
+        Node.main_node.send_data_to_node(each_node, items)
